requirements.txt
# Існуючі залежності
python-dotenv>=1.0.0
pdfplumber>=0.10.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Preprocessing
marker-pdf>=0.2.0
nltk>=3.8.0
regex>=2023.0.0
python-docx>=1.1.0

# LangChain (нові)
langchain>=1.0.0
langchain-community
langchain-ollama
langgraph
langserve

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0

# Optional
langsmith>=0.1.0
